refactor: remove all builders to VersionManagerBuilders class
