+++ packages/flutter_version_manager/lib/flutter_version_manager.dart
@@ -1,6 +1,7 @@
 export "package:version_repository_interface/version_repository_interface.dart";
 
 export "src/check_updates.dart";
+export "src/config/builders.dart";
 export "src/config/translations.dart";
 export "src/config/version_manager_config.dart";
 export "src/flutter_version_manager.dart";
